fix: rename distribution to hola-opt, replace --legacy with --mode, clean up docs
- Rename PyPI distribution from "hola" to "hola-opt" and Python
  module from "hola" to "hola_opt" to avoid collision with an
  unrelated package on PyPI.
- Replace --legacy CLI flag with --mode callback|exec.
- Update README to describe callback mode (the default) instead
  of the stdout-parsing workflow.
- Reconcile study.yaml vs config.yaml naming in README/CLI docs.
- Bump actions/checkout from v4 to v5 in all CI workflows.
- Update PEP 503 index paths in release workflow for hola-opt.
