+++ hola-py/examples/gmm_explore_exploit.py
@@ -23,7 +23,7 @@ The GMM strategy:
 """
 
 from benchmarks.functions.single_objective import branin, rastrigin
-from hola import Minimize, Real, Space, Study
+from hola_opt import Minimize, Real, Space, Study
 
 
 def _make_nd_bounds(n: int, lo: float, hi: float) -> dict[str, tuple[float, float]]:
